Se crea archivo de logUtils modularizado y aplicado en las distintas partes del programa
